Confirm on all deleting, some formating

+++ Momiji/Forms/frmGSManager.cs
@@ -208,6 +208,9 @@ namespace Momiji
 
 		protected void OnBtnDeleteClicked (object sender, EventArgs e)
 		{
+			if (!MessageBox.Ask (this, "Are you sure you want to delete this piece?"))
+				return;
+
 			SQL SQLConnection = parent.currentSQLConnection;
 			MySqlCommand query = new MySqlCommand ("DELETE FROM `gsmerchandise` WHERE `ArtistID` = @AID AND `PieceID` = @PID;",
 				                     SQLConnection.GetConnection ());
